Cleaned up imports and redundant functions.

+++ TrainingAPI/src/main/java/org/demo/repositories/TrainingRepository.java
@@ -1,8 +1,7 @@
 package org.demo.repositories;
 
-import org.demo.model.TrainingPOJO;
+import org.demo.entity.TrainingPOJO;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
-public interface TrainingRepository extends JpaRepository<TrainingPOJO,Long> {
+public interface TrainingRepository extends JpaRepository<TrainingPOJO,Integer> {
 }
